Add Beeman class & clean code
