--- src/main/java/ar/edu/itba/ss/tpf/Mode.java
@@ -1,22 +1,0 @@
-package ar.edu.itba.ss.tpf;
-
-import java.util.Arrays;
-import java.util.Optional;
-
-public enum Mode {
-	SINGLE_RUN (0);
-	
-	private int mode;
-	
-	Mode(int mode) {
-		this.mode = mode;
-	}
-	
-	public int getMode() {
-		return mode;
-	}
-	
-	public static Optional<Mode> valueOf(int value) {
-		return Arrays.stream(values()).filter(m -> m.getMode() == value).findFirst();
-	}
-}
